refactor: split out Disco.Value and Disco.Error from Disco.Eval
