+++ src/Disco/Interactive/Commands.hs
@@ -54,6 +54,8 @@ import           Disco.Typecheck
 import           Disco.Typecheck.Erase
 import           Disco.Typecheck.Monad
 import           Disco.Types
+import           Disco.Util
+import           Disco.Value
 import           Text.Megaparsec                  hiding (runParser)
 import           Unbound.Generics.LocallyNameless
